`windows-gnu` -> `windows-mingw`
`windows-gnu` has been used by LLVM and Rust to mean MinGW, but many
people on the mailing list object that it is in no way bona fide GNU on
Windows. Even under the interpretation of `-gnu` as a libc which LLVM
often goes with, it doesn't pass muster either because MinGW uses
msvcrt/ucrt, not any GNU libc. Arguably Cygwin, using a modified (GNU)
Newlib, is the closest thing we have to "GNU on Windows" today.

"MinGW" stands for "Minimalist GNU for Windows", and so `windows-mingw`
redundantly both has "windows" and the second "w". However "mingw" is a
well-known years old name at this point, and there certainly could be
other "Minimalist GNU"s, so its best to treat the contraction as a black
box rather than try to decompose it into its constituent pieces.