`windows-gnu` -> `windows-mingnu`
`windows-gnu` has been used by LLVM and Rust to mean MinGW, but many
people on the mailing list object that it is in no way bona fide GNU on
Windows. Even under the interpretation of `-gnu` as a libc which LLVM
often goes with, it doesn't pass muster either because MinGW uses
msvcrt/ucrt, not any GNU libc. Arguably Cygwin, using a modified (GNU)
Newlib, is the closest thing we have to "GNU on Windows" today.

`mingnu` is not yet in use, but the idea is "MinGW" stands for
"Minimalist GNU for Windows", and so `windows-mingnu` refers to all
three main words of that name.